Updating README with some guesses
The way Clerk is used has changed drastically. This is a first
stab at getting the README updated to reflect the new Templating
which also led to some tweaks for the rest of the usage of Clerk.
The sections about how to use parsing will have to be fleshed out
elsewhere.